pyproject.toml
[project]
name = "cabernet"
version = "0.1.0"
description = ""
authors = [
    {name = "V.Guilherme",email = "v.guilherme.barreto@gmail.com"}
]
readme = "README.md"
requires-python = "3.13.5"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "Flask-JWT-Extended (>=4.6.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
dev = 'poetry run flask --app cabernet/app run --debug'
db = 'poetry run flask --app cabernet/app db'
clean = 'find . -type d -name "__pycache__" -exec rm -rf {} + && find . -type f -name "*.pyc" -delete'
